requirements.txt
django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
google-api-python-client==2.154.0
openai==2.21.0
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.7.0
